Move IrohConnection to peer.rs, to keep all Iroh code in one file
